+++ src/main/java/com/codestates/pre012/posts/entity/Posts.java
@@ -4,7 +4,7 @@ package com.codestates.pre012.posts.entity;
 import com.codestates.pre012.baseEntity.BaseEntity;
 import com.codestates.pre012.member.entity.Member;
 import com.codestates.pre012.reply.Reply;
-import com.codestates.pre012.tag.entity.Tag_Posts;
+import com.codestates.pre012.tag.Tag_Posts;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
