+++ src/main/java/com/codestates/pre012/tag/TagController.java
@@ -1,4 +1,4 @@
-package com.codestates.pre012.tag.controller;
+package com.codestates.pre012.tag;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
