+++ src/main/java/com/codestates/pre012/tag/TagDto.java
@@ -1,13 +1,9 @@
-package com.codestates.pre012.tag.dto;
+package com.codestates.pre012.tag;
 
-import com.codestates.pre012.tag.entity.Tag;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Positive;
 
 public class TagDto {
     @Getter
