+++ src/main/java/com/codestates/pre012/tag/TagMapper.java
@@ -1,7 +1,5 @@
 package com.codestates.pre012.tag;
 
-import com.codestates.pre012.tag.dto.TagDto;
-import com.codestates.pre012.tag.entity.Tag;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
