+++ src/main/java/com/codestates/pre012/tag/TagRepository.java
@@ -1,9 +1,7 @@
-package com.codestates.pre012.tag.repository;
+package com.codestates.pre012.tag;
 
-import com.codestates.pre012.tag.entity.Tag;
+import com.codestates.pre012.tag.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
