+++ src/main/java/com/codestates/pre012/tag/TagService.java
@@ -1,0 +1,16 @@
+package com.codestates.pre012.tag;
+
+import com.codestates.pre012.tag.TagRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class TagService {
+
+    private final TagRepository tagRepository;
+
+
+
+
+}
