+++ src/main/java/com/codestates/pre012/tag/Tag_PostsDto.java
@@ -1,0 +1,4 @@
+package com.codestates.pre012.tag;
+
+public class Tag_PostsDto {
+}
