+++ src/main/java/com/codestates/pre012/tag/Tag_PostsRepository.java
@@ -1,6 +1,6 @@
-package com.codestates.pre012.tag.repository;
+package com.codestates.pre012.tag;
 
-import com.codestates.pre012.tag.entity.Tag_Posts;
+import com.codestates.pre012.tag.Tag_Posts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface Tag_PostsRepository extends JpaRepository<Tag_Posts, Long> {
