--- src/main/java/com/codestates/pre012/tag/dto/Tag_PostsDto.java
@@ -1,4 +1,0 @@
-package com.codestates.pre012.tag.dto;
-
-public class Tag_PostsDto {
-}
