+++ src/main/resources/application.yml
@@ -4,13 +4,13 @@ spring:
       enabled: true
       path: /h2
   datasource:
-#    url: jdbc:mysql://admin.cy6doy34twwm.ap-northeast-2.rds.amazonaws.com/team012?useSSL=false&characterEncoding=UTF-8&serverTimezone=Asia/Seoul
-#    username: admin
-#    password:
-#    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://admin.cy6doy34twwm.ap-northeast-2.rds.amazonaws.com/team012?useSSL=false&characterEncoding=UTF-8&serverTimezone=Asia/Seoul
+    username: admin
+    password:
+    driver-class-name: com.mysql.cj.jdbc.Driver
 
 
-      url: jdbc:h2:mem:test  # h2
+#      url: jdbc:h2:mem:test  # h2
 #  admin.cy6doy34twwm.ap-northeast-2.rds.amazonaws.com
 
   security:
